managed to get composite models working on routes
the key is that the todos/index route is the one that is displayed, and
is the only one that can actually render templates; I am guessing that
routes do not call parents except to mess things up. so by moving the
renderTemplates into the child todos/index route, we can render the two
outlets separately (but I guess it means we have to duplicate the
renderTemplates for each outlet on each route/view).